LayoutTests/platform/gtk/fast/dom/Range/getClientRects-expected.txt

Test 1
PASS rects.length is 5
PASS rects[0].left.toFixed(2) is "8.00"
PASS rects[0].top.toFixed(2) is "8.00"
PASS rects[0].width.toFixed(2) is "400.00"
PASS rects[0].height.toFixed(2) is "160.00"
PASS rects[1].left.toFixed(2) is "8.00"
PASS rects[1].top.toFixed(2) is "19.00"
FAIL rects[1].width.toFixed(2) should be 386.58. Was 396.00.
FAIL rects[1].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[2].left.toFixed(2) is "8.00"
PASS rects[2].top.toFixed(2) is "59.00"
FAIL rects[2].width.toFixed(2) should be 359.91. Was 398.00.
FAIL rects[2].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[3].left.toFixed(2) is "8.00"
PASS rects[3].top.toFixed(2) is "99.00"
FAIL rects[3].width.toFixed(2) should be 381.25. Was 360.00.
FAIL rects[3].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[4].left.toFixed(2) is "8.00"
PASS rects[4].top.toFixed(2) is "139.00"
FAIL rects[4].width.toFixed(2) should be 368.36. Was 306.00.
FAIL rects[4].height.toFixed(2) should be 18.00. Was 17.00.
Test 2
PASS rects.length is 4
PASS rects[0].left.toFixed(2) is "8.00"
FAIL rects[0].top.toFixed(2) should be 215.00. Was 213.00.
FAIL rects[0].width.toFixed(2) should be 386.58. Was 396.00.
FAIL rects[0].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[1].left.toFixed(2) is "8.00"
FAIL rects[1].top.toFixed(2) should be 255.00. Was 253.00.
FAIL rects[1].width.toFixed(2) should be 359.91. Was 398.00.
FAIL rects[1].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[2].left.toFixed(2) is "8.00"
FAIL rects[2].top.toFixed(2) should be 295.00. Was 293.00.
FAIL rects[2].width.toFixed(2) should be 381.25. Was 360.00.
FAIL rects[2].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[3].left.toFixed(2) is "8.00"
FAIL rects[3].top.toFixed(2) should be 335.00. Was 333.00.
FAIL rects[3].width.toFixed(2) should be 368.36. Was 306.00.
FAIL rects[3].height.toFixed(2) should be 18.00. Was 17.00.
Test 3
PASS rects.length is 4
FAIL rects[0].left.toFixed(2) should be 81.31. Was 80.00.
FAIL rects[0].top.toFixed(2) should be 411.00. Was 407.00.
FAIL rects[0].width.toFixed(2) should be 313.27. Was 324.00.
FAIL rects[0].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[1].left.toFixed(2) is "8.00"
FAIL rects[1].top.toFixed(2) should be 451.00. Was 447.00.
FAIL rects[1].width.toFixed(2) should be 359.91. Was 398.00.
FAIL rects[1].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[2].left.toFixed(2) is "8.00"
FAIL rects[2].top.toFixed(2) should be 491.00. Was 487.00.
FAIL rects[2].width.toFixed(2) should be 381.25. Was 360.00.
FAIL rects[2].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[3].left.toFixed(2) is "8.00"
FAIL rects[3].top.toFixed(2) should be 531.00. Was 527.00.
FAIL rects[3].width.toFixed(2) should be 146.66. Was 88.00.
FAIL rects[3].height.toFixed(2) should be 18.00. Was 17.00.
Test 4
PASS rects.length is 7
PASS rects[0].left.toFixed(2) is "8.00"
FAIL rects[0].top.toFixed(2) should be 596.00. Was 590.00.
PASS rects[0].width.toFixed(2) is "400.00"
PASS rects[0].height.toFixed(2) is "278.00"
PASS rects[1].left.toFixed(2) is "8.00"
FAIL rects[1].top.toFixed(2) should be 607.00. Was 601.00.
FAIL rects[1].width.toFixed(2) should be 386.58. Was 396.00.
FAIL rects[1].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[2].left.toFixed(2) is "8.00"
FAIL rects[2].top.toFixed(2) should be 725.00. Was 719.00.
FAIL rects[2].width.toFixed(2) should be 268.39. Was 242.00.
FAIL rects[2].height.toFixed(2) should be 18.00. Was 17.00.
FAIL rects[3].left.toFixed(2) should be 352.39. Was 326.00.
FAIL rects[3].top.toFixed(2) should be 725.00. Was 719.00.
FAIL rects[3].width.toFixed(2) should be 44.88. Was 44.00.
FAIL rects[3].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[4].left.toFixed(2) is "8.00"
FAIL rects[4].top.toFixed(2) should be 765.00. Was 759.00.
FAIL rects[4].width.toFixed(2) should be 348.38. Was 341.00.
FAIL rects[4].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[5].left.toFixed(2) is "8.00"
FAIL rects[5].top.toFixed(2) should be 805.00. Was 799.00.
FAIL rects[5].width.toFixed(2) should be 375.90. Was 366.00.
FAIL rects[5].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[6].left.toFixed(2) is "8.00"
FAIL rects[6].top.toFixed(2) should be 845.00. Was 839.00.
FAIL rects[6].width.toFixed(2) should be 67.98. Was 67.00.
FAIL rects[6].height.toFixed(2) should be 18.00. Was 17.00.
Test 5
PASS rects.length is 5
FAIL rects[0].left.toFixed(2) should be 81.00. Was 80.00.
FAIL rects[0].top.toFixed(2) should be 921.00. Was 913.00.
FAIL rects[0].width.toFixed(2) should be 313.58. Was 324.00.
FAIL rects[0].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[1].left.toFixed(2) is "8.00"
FAIL rects[1].top.toFixed(2) should be 1039.00. Was 1031.00.
FAIL rects[1].width.toFixed(2) should be 268.39. Was 242.00.
FAIL rects[1].height.toFixed(2) should be 18.00. Was 17.00.
FAIL rects[2].left.toFixed(2) should be 276.39. Was 250.00.
FAIL rects[2].top.toFixed(2) should be 950.00. Was 942.00.
PASS rects[2].width.toFixed(2) is "76.00"
PASS rects[2].height.toFixed(2) is "103.00"
FAIL rects[3].left.toFixed(2) should be 352.39. Was 326.00.
FAIL rects[3].top.toFixed(2) should be 1039.00. Was 1031.00.
FAIL rects[3].width.toFixed(2) should be 44.88. Was 44.00.
FAIL rects[3].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[4].left.toFixed(2) is "8.00"
FAIL rects[4].top.toFixed(2) should be 1079.00. Was 1071.00.
FAIL rects[4].width.toFixed(2) should be 291.00. Was 284.00.
FAIL rects[4].height.toFixed(2) should be 18.00. Was 17.00.
Test 6
FAIL rects.length should be 8. Was 7.
PASS rects[0].left.toFixed(2) is "8.00"
FAIL rects[0].top.toFixed(2) should be 1224.00. Was 1214.00.
PASS rects[0].width.toFixed(2) is "400.00"
FAIL rects[0].height.toFixed(2) should be 240.00. Was 200.00.
PASS rects[1].left.toFixed(2) is "8.00"
FAIL rects[1].top.toFixed(2) should be 1235.00. Was 1225.00.
FAIL rects[1].width.toFixed(2) should be 386.58. Was 396.00.
FAIL rects[1].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[2].left.toFixed(2) is "8.00"
FAIL rects[2].top.toFixed(2) should be 1275.00. Was 1265.00.
FAIL rects[2].width.toFixed(2) should be 16.00. Was 57.00.
FAIL rects[2].height.toFixed(2) should be 18.00. Was 17.00.
FAIL rects[3].left.toFixed(2) should be 8.00. Was 105.00.
FAIL rects[3].top.toFixed(2) should be 1315.00. Was 1265.00.
FAIL rects[3].width.toFixed(2) should be 58.22. Was 44.00.
FAIL rects[3].height.toFixed(2) should be 18.00. Was 17.00.
FAIL rects[4].left.toFixed(2) should be 106.22. Was 8.00.
FAIL rects[4].top.toFixed(2) should be 1315.00. Was 1305.00.
FAIL rects[4].width.toFixed(2) should be 45.32. Was 387.00.
FAIL rects[4].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[5].left.toFixed(2) is "8.00"
FAIL rects[5].top.toFixed(2) should be 1355.00. Was 1345.00.
FAIL rects[5].width.toFixed(2) should be 396.34. Was 397.00.
FAIL rects[5].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[6].left.toFixed(2) is "8.00"
FAIL rects[6].top.toFixed(2) should be 1395.00. Was 1385.00.
FAIL rects[6].width.toFixed(2) should be 359.93. Was 171.00.
FAIL rects[6].height.toFixed(2) should be 18.00. Was 17.00.
FAIL rects[7].left.toFixed(2) should be 8.00. Threw exception TypeError: undefined is not an object (evaluating 'rects[7].left')
FAIL rects[7].top.toFixed(2) should be 1435.00. Threw exception TypeError: undefined is not an object (evaluating 'rects[7].top')
FAIL rects[7].width.toFixed(2) should be 221.71. Threw exception TypeError: undefined is not an object (evaluating 'rects[7].width')
FAIL rects[7].height.toFixed(2) should be 18.00. Threw exception TypeError: undefined is not an object (evaluating 'rects[7].height')
Test 7
FAIL rects.length should be 8. Was 7.
PASS rects[0].left.toFixed(2) is "8.00"
FAIL rects[0].top.toFixed(2) should be 1511.00. Was 1459.00.
FAIL rects[0].width.toFixed(2) should be 386.58. Was 396.00.
FAIL rects[0].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[1].left.toFixed(2) is "8.00"
FAIL rects[1].top.toFixed(2) should be 1551.00. Was 1488.00.
FAIL rects[1].width.toFixed(2) should be 16.00. Was 400.00.
FAIL rects[1].height.toFixed(2) should be 18.00. Was 40.00.
PASS rects[2].left.toFixed(2) is "8.00"
FAIL rects[2].top.toFixed(2) should be 1580.00. Was 1499.00.
FAIL rects[2].width.toFixed(2) should be 400.00. Was 57.00.
FAIL rects[2].height.toFixed(2) should be 40.00. Was 17.00.
FAIL rects[3].left.toFixed(2) should be 8.00. Was 105.00.
FAIL rects[3].top.toFixed(2) should be 1591.00. Was 1499.00.
FAIL rects[3].width.toFixed(2) should be 58.22. Was 44.00.
FAIL rects[3].height.toFixed(2) should be 18.00. Was 17.00.
FAIL rects[4].left.toFixed(2) should be 106.22. Was 8.00.
FAIL rects[4].top.toFixed(2) should be 1591.00. Was 1539.00.
FAIL rects[4].width.toFixed(2) should be 45.32. Was 387.00.
FAIL rects[4].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[5].left.toFixed(2) is "8.00"
FAIL rects[5].top.toFixed(2) should be 1631.00. Was 1579.00.
FAIL rects[5].width.toFixed(2) should be 396.34. Was 397.00.
FAIL rects[5].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[6].left.toFixed(2) is "8.00"
FAIL rects[6].top.toFixed(2) should be 1671.00. Was 1619.00.
FAIL rects[6].width.toFixed(2) should be 359.93. Was 171.00.
FAIL rects[6].height.toFixed(2) should be 18.00. Was 17.00.
FAIL rects[7].left.toFixed(2) should be 8.00. Threw exception TypeError: undefined is not an object (evaluating 'rects[7].left')
FAIL rects[7].top.toFixed(2) should be 1711.00. Threw exception TypeError: undefined is not an object (evaluating 'rects[7].top')
FAIL rects[7].width.toFixed(2) should be 221.71. Threw exception TypeError: undefined is not an object (evaluating 'rects[7].width')
FAIL rects[7].height.toFixed(2) should be 18.00. Threw exception TypeError: undefined is not an object (evaluating 'rects[7].height')
Test 8
PASS rects.length is 4
FAIL rects[0].left.toFixed(2) should be 168.51. Was 169.28.
FAIL rects[0].top.toFixed(2) should be 1758.44. Was 1664.44.
FAIL rects[0].width.toFixed(2) should be 262.28. Was 267.57.
FAIL rects[0].height.toFixed(2) should be 307.71. Was 314.28.
FAIL rects[1].left.toFixed(2) should be 137.87. Was 138.64.
FAIL rects[1].top.toFixed(2) should be 1784.15. Was 1690.15.
FAIL rects[1].width.toFixed(2) should be 245.14. Was 268.85.
FAIL rects[1].height.toFixed(2) should be 287.28. Was 315.81.
FAIL rects[2].left.toFixed(2) should be 107.23. Was 107.99.
FAIL rects[2].top.toFixed(2) should be 1809.86. Was 1715.86.
FAIL rects[2].width.toFixed(2) should be 258.85. Was 244.43.
FAIL rects[2].height.toFixed(2) should be 303.62. Was 286.70.
FAIL rects[3].left.toFixed(2) should be 76.59. Was 77.35.
FAIL rects[3].top.toFixed(2) should be 1835.57. Was 1741.57.
FAIL rects[3].width.toFixed(2) should be 250.57. Was 209.72.
FAIL rects[3].height.toFixed(2) should be 293.75. Was 245.34.
Test 9
PASS rects.length is 0
Test 9b
PASS rects.length is 1
PASS rects[0].left.toFixed(2) is "8.00"
FAIL rects[0].top.toFixed(2) should be 1983.00. Was 1887.00.
PASS rects[0].width.toFixed(2) is "0.00"
FAIL rects[0].height.toFixed(2) should be 18.00. Was 17.00.
Test 10
FAIL rects.length should be 5. Was 4.
PASS rects[0].left.toFixed(2) is "19.00"
FAIL rects[0].top.toFixed(2) should be 2168.00. Was 2070.00.
FAIL rects[0].width.toFixed(2) should be 18.00. Was 17.00.
FAIL rects[0].height.toFixed(2) should be 375.00. Was 393.00.
PASS rects[1].left.toFixed(2) is "59.00"
FAIL rects[1].top.toFixed(2) should be 2168.00. Was 2070.00.
FAIL rects[1].width.toFixed(2) should be 18.00. Was 17.00.
FAIL rects[1].height.toFixed(2) should be 367.03. Was 377.00.
PASS rects[2].left.toFixed(2) is "99.00"
FAIL rects[2].top.toFixed(2) should be 2168.00. Was 2070.00.
FAIL rects[2].width.toFixed(2) should be 18.00. Was 17.00.
FAIL rects[2].height.toFixed(2) should be 348.38. Was 372.00.
PASS rects[3].left.toFixed(2) is "139.00"
FAIL rects[3].top.toFixed(2) should be 2168.00. Was 2070.00.
FAIL rects[3].width.toFixed(2) should be 18.00. Was 17.00.
FAIL rects[3].height.toFixed(2) should be 375.90. Was 360.00.
FAIL rects[4].left.toFixed(2) should be 179.00. Threw exception TypeError: undefined is not an object (evaluating 'rects[4].left')
FAIL rects[4].top.toFixed(2) should be 2168.00. Threw exception TypeError: undefined is not an object (evaluating 'rects[4].top')
FAIL rects[4].width.toFixed(2) should be 18.00. Threw exception TypeError: undefined is not an object (evaluating 'rects[4].width')
FAIL rects[4].height.toFixed(2) should be 67.98. Threw exception TypeError: undefined is not an object (evaluating 'rects[4].height')
Test 11
PASS rects.length is 2
PASS rects[0].left.toFixed(2) is "8.00"
FAIL rects[0].top.toFixed(2) should be 2604.00. Was 2504.00.
PASS rects[0].width.toFixed(2) is "400.00"
PASS rects[0].height.toFixed(2) is "40.00"
PASS rects[1].left.toFixed(2) is "8.00"
FAIL rects[1].top.toFixed(2) should be 2615.00. Was 2515.00.
FAIL rects[1].width.toFixed(2) should be 181.75. Was 177.00.
FAIL rects[1].height.toFixed(2) should be 18.00. Was 17.00.
Test 12
PASS rects.length is 1
PASS rects[0].left.toFixed(2) is "8.00"
FAIL rects[0].top.toFixed(2) should be 2840.00. Was 2738.00.
PASS rects[0].width.toFixed(2) is "400.00"
PASS rects[0].height.toFixed(2) is "160.00"
Test 13
PASS rects.length is 4
PASS rects[0].left.toFixed(2) is "8.00"
FAIL rects[0].top.toFixed(2) should be 3047.00. Was 2943.00.
FAIL rects[0].width.toFixed(2) should be 386.58. Was 396.00.
FAIL rects[0].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[1].left.toFixed(2) is "8.00"
FAIL rects[1].top.toFixed(2) should be 3087.00. Was 2983.00.
FAIL rects[1].width.toFixed(2) should be 359.91. Was 398.00.
FAIL rects[1].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[2].left.toFixed(2) is "8.00"
FAIL rects[2].top.toFixed(2) should be 3127.00. Was 3023.00.
FAIL rects[2].width.toFixed(2) should be 381.25. Was 360.00.
FAIL rects[2].height.toFixed(2) should be 18.00. Was 17.00.
PASS rects[3].left.toFixed(2) is "8.00"
FAIL rects[3].top.toFixed(2) should be 3167.00. Was 3063.00.
FAIL rects[3].width.toFixed(2) should be 368.36. Was 306.00.
FAIL rects[3].height.toFixed(2) should be 18.00. Was 17.00.
Test 14
PASS rects.length is 0
PASS successfullyParsed is true

TEST COMPLETE
